:recycle: aws: restructure serverless config directory
